+++ src/api/delete_image.rs
@@ -1,5 +1,6 @@
-use super::send_api_request;
 use crate::api::configuration::{api_url, ImgurHandle};
+
+use super::send_api_request;
 
 use log::error;
 use reqwest::Method;