+++ src/api/upload_image.rs
@@ -1,8 +1,9 @@
-use super::send_api_request;
 use crate::api::{
     configuration::{api_url, ImgurHandle},
     ImageInfo,
 };
+
+use super::send_api_request;
 
 use log::error;
 use reqwest::Method;