+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased]
+
+## [0.3.0] - 2022-01-28
 ### CLI
 - SimpleLogger init error handling
 - better panic
@@ -50,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - upload image
 
 <!-- next-url -->
-[Unreleased]: https://github.com/MedzikUser/imgurs/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/MedzikUser/imgurs/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/MedzikUser/imgurs/commits/v0.3.0
 [0.2.0]: https://github.com/MedzikUser/imgurs/commits/v0.2.0
 [0.1.0]: https://github.com/MedzikUser/imgurs/commits/v0.1.0